% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{K-nearest neighbors cross-validation}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Data set to run model on}

\item{cl}{Values for the true class}

\item{k_nn}{Number of closest neighbors}

\item{k_cv}{Total number of folds per model}
}
\value{
Output list with predicted class, cross-validation error,
and training error
}
\description{
This function executes a knn model based on input data including
cross-validation and training error results
}
\examples{
my_knn_cv(new_penguins[, -c(1, 2, 7, 8)], new_penguins$species, 1, 5)
my_knn_cv(new_penguins[, -c(1, 2, 7, 8)], new_penguins$species, 5, 5)


}
\keyword{prediction}
