% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{Linear Regression Model}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{Formula to run data correlations}

\item{data}{Dataset which contained inputs run through formula}
}
\value{
Data frame containing values for standard errors, betas,
t-statistics, and p-values
}
\description{
This function creates a linear regression model based on data input
}
\examples{
data("my_penguins")
new_penguins <- na.omit(my_penguins)
my_lm(lifeExp ~ gdpPercap + continent, my_gapminder)
my_lm(body_mass_g ~ flipper_length_mm + island, new_penguins)

}
\keyword{inference}
