% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{Random Forest Cross-Validation}
\usage{
my_rf_cv(train, k)
}
\arguments{
\item{train}{Data set to run model on}

\item{k}{Total number of folds per model}
}
\value{
Output list with values of the average MSE from total folds
}
\description{
This function runs a random forest cross-validation prediction
for a given set of data and calculates the the MSE errors
}
\examples{
data("my_penguins")
new_penguins <- na.omit(my_penguins)
my_rf_cv(new_penguins, 5)
my_rf_cv(new_penguins, 10)

}
\keyword{prediction}
