% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_test.R
\name{my_test}
\alias{my_test}
\title{T.Test Function}
\usage{
my_test(x, alternative, mu)
}
\arguments{
\item{x}{Dataset on which t.test will operate}

\item{alternative}{Strings of either "two.sided", "lesser", or "greater" to
designate alternative hypothesis}

\item{mu}{Numeric input designating value of the null hypothesis}
}
\value{
Output list with dataset's standard error, test statistic, degrees
freedom, alternative hypothesis, and p-value after run through t.test
}
\description{
This function conducts a t.test with a given dataset
}
\examples{
my_test(my_gapminder$lifeExp, "two.sided", 60)
my_test(1:100, "two.sided", 60)

}
\keyword{inference}
